Change binder link to AARNet service.
Alex Ip
  下午 1:22
ANNOUNCEMENT: Binder is finally (mostly) working on the NECTAR cloud (http://203.100.31.148/), thanks to
@River
's help.
It's not in its final state yet, but we now have it building an image from a GitHub repo, and then pushing it to DockerHub.
Note that the initial run doesn't go on to launch the Jupyter notebook from the image (possibly a timeout), but a subsequent call to the binder for the same repo does. A simple page refresh also works, and gives a redirect to the Jupyter notebook. （已编辑）